refactor(*): Move Environment to lang
